cityscope_v3/backend/issues/permissions.py
from rest_framework.permissions import BasePermission


class IsAdminUser(BasePermission):
    """Only allow staff/admin users."""
    message = 'Admin access required.'

    def has_permission(self, request, view):
        return bool(
            request.user
            and request.user.is_authenticated
            and request.user.is_staff
        )
