cityscope_v3/backend/requirements.txt
Django==4.1.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.3.0
python-decouple==3.8
mysqlclient==2.2.4
